Fix: estruturação
