+++ projeto18-linkr-frontend/src/API_URL.js
@@ -1,0 +1,1 @@
+export const API_URL = 'localhost:5000/'
