--- projeto18-linkr-frontend/src/components/API_URL.js
@@ -1,1 +1,0 @@
-export const API_URL = ''
