+++ projeto18-linkr-frontend/src/components/button/Button.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+export default function Button() {
+    return(<>
+    <ButtonExemple></ButtonExemple>
+    </>)
+}
+const ButtonExemple = styled.div`
+
+`
