+++ projeto18-linkr-frontend/src/components/input/Input.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+export default function Input() {
+    return(<>
+    <InputExemple></InputExemple>
+    </>)
+}
+
+const InputExemple = styled.input`
+
+`
