+++ projeto18-linkr-frontend/src/index.js
@@ -1,4 +1,4 @@
 import ReactDOM from "react-dom";
-import App from "./components/App";
+import App from "./App";
 
 ReactDOM.render(<App />, document.querySelector(".root"));
